backend/pyproject.toml
[tool.pdm]

[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Heng", email = "31088159+shiukaheng@users.noreply.github.com"},
]
dependencies = [
    "fastapi>=0.92.0",
    "uvicorn>=0.20.0",
    "mechanicalsoup>=1.2.0",
    "openai>=0.27.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
